# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: nuruzzaman.com.bd API
  version: 1.0.0
  description: |
    First-party API for nuruzzaman.com.bd.

    Authentication is a Laravel Sanctum cookie session, not a bearer token: the
    browser calls `GET /sanctum/csrf-cookie` once, then sends the `XSRF-TOKEN`
    value back in the `X-XSRF-TOKEN` header on every mutating request. Nothing
    is stored in `localStorage`.

    Money is always an integer in minor units (1 BDT = 100 poisha). A `null`
    price means the owner has not published one yet and the UI must render an
    honest "contact for price" state rather than zero.

    Every error uses the same envelope, so one client-side handler covers the
    whole surface.
  contact:
    name: Engr. Md. Nuruzzaman, RSE
  license:
    name: Proprietary
servers:
  - url: https://nuruzzaman.com.bd/api/v1
    description: Production (same origin, proxied to Laravel by Nginx)
  - url: http://localhost:8000/api/v1
    description: Local development
...
security: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nuruzzaman_session
      description: First-party Sanctum session cookie set by /api/v1/auth/login.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Makes a payment-sensitive mutation safe to retry. Repeating the same key
        with the same body replays the stored response; reusing it with a
        different body is a 409.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Unauthenticated:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: The request conflicts with the current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationFailed:
      description: Field-level validation errors
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation_failed
                - unauthenticated
                - forbidden
                - not_found
                - conflict
                - gone
                - locked
                - rate_limited
                - bad_request
                - request_failed
                - server_error
            message: { type: string }
            fields:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
            request_id: { type: string }
...
    OrderItem:
      type: object
      required: [product_name, variant_name, sku, product_type, quantity, unit_price_minor, line_total_minor]
      properties:
        product_name: { type: string }
        variant_name: { type: string }
        sku: { type: string }
        product_type: { type: string }
        quantity: { type: integer }
        unit_price_minor: { type: integer }
        line_total_minor: { type: integer }
    Order:
      type: object
      required: [number, status, currency, subtotal_minor, discount_minor, tax_minor, total_minor, refunded_minor, billing_name, billing_email, placed_at, paid_at, fulfilled_at, accepted_terms]
      properties:
        number: { type: string }
        status:
          type: string
          enum:
            - draft
            - pending_payment
            - paid
            - fulfilled
            - failed
            - cancelled
            - refund_pending
            - partially_refunded
            - refunded
        currency: { type: string }
        subtotal_minor: { type: integer }
        discount_minor: { type: integer }
        tax_minor: { type: integer }
        total_minor: { type: integer }
        refunded_minor: { type: integer }
        billing_name: { type: [string, 'null'] }
        billing_email: { type: [string, 'null'] }
        placed_at: { type: [string, 'null'], format: date-time }
        paid_at: { type: [string, 'null'], format: date-time }
        fulfilled_at: { type: [string, 'null'], format: date-time }
        accepted_terms: { type: array, items: { type: string } }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        invoice_number: { type: [string, 'null'] }
        timeline:
          type: array
          items:
            type: object
            required: [from, to, reason, at]
            properties:
              from: { type: [string, 'null'] }
              to: { type: string }
              reason: { type: [string, 'null'] }
              at: { type: [string, 'null'], format: date-time }
...
paths:
...
  /checkout:
    post:
      tags: [Checkout]
      operationId: checkout
      summary: Checkout
      security: [{ sessionCookie: [] }]
      description: |
        Creates a pending order from the server-priced cart and opens a gateway
        session. Any total sent by the browser is ignored.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, accepts_terms, accepts_privacy, accepts_refund_policy]
              properties:
                name: { type: string, minLength: 2, maxLength: 160 }
                email: { type: string, format: email }
                phone: { type: [string, 'null'] }
                accepts_terms: { type: boolean, enum: [true] }
                accepts_privacy: { type: boolean, enum: [true] }
                accepts_refund_policy: { type: boolean, enum: [true] }
                accepts_eula: { type: boolean }
      responses:
        '201':
          description: Order created; send the customer to redirect_url
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [order, payment_reference, redirect_url]
                    properties:
                      order: { $ref: '#/components/schemas/Order' }
                      payment_reference: { type: [string, 'null'] }
                      redirect_url: { type: string, format: uri }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503':
          description: The payment gateway is not configured or not reachable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...